working on QuestionViewModel

+++ app/src/main/java/biz/eventually/atpl/data/network/AnswerNetwork.kt
@@ -3,4 +3,4 @@ package biz.eventually.atpl.data.network
 /**
  * Created by thibault on 21/03/17.
  */
-data class AnswerNetwork(val id: Int, val value: String, val good: Boolean = false)
+data class AnswerNetwork(val id: Long, val value: String, val good: Boolean = false)
